requirements.txt
pydantic==2.13.1
uvicorn==0.44.0
pytesseract==0.3.13
pillow==12.2.0
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc